refactor(extension): barrel file(index.ts) 적용으로 import 구조 개선
