+++ apps/extension/src/components/index.ts
@@ -1,0 +1,6 @@
+export { default as DatePicker } from "./DatePicker";
+export { default as Divider } from "./Divider";
+export * from "./GNBTabs";
+export { default as Icon } from "./Icon";
+export { default as ScreenTimeWeeklyBarChart } from "./ScreenTimeWeeklyBarChart";
+export * from "./ToggleGroup";
