+++ apps/extension/src/features/ai-recap/components/index.ts
@@ -1,0 +1,1 @@
+export { default as AiRecapView } from "./AiRecapView";
