+++ apps/extension/src/features/analysis/components/AnalysisViewSkeleton.tsx
@@ -1,7 +1,8 @@
-import Divider from "@/components/Divider";
-import CategoryAnalysisSectionSkeleton from "@/features/analysis/components/CategoryAnalysisSectionSkeleton";
-import TodayTimeThiefSectionSkeleton from "@/features/analysis/components/TodayTimeThiefSectionSkeleton";
-import WeeklyScreenTimeSectionSkeleton from "@/features/analysis/components/WeeklyScreenTimeSectionSkeleton";
+import { Divider } from "@/components";
+
+import CategoryAnalysisSectionSkeleton from "./CategoryAnalysisSectionSkeleton";
+import TodayTimeThiefSectionSkeleton from "./TodayTimeThiefSectionSkeleton";
+import WeeklyScreenTimeSectionSkeleton from "./WeeklyScreenTimeSectionSkeleton";
 
 const AnalysisViewSkeleton = () => {
   return (
