+++ apps/extension/src/features/analysis/components/index.ts
@@ -1,0 +1,1 @@
+export { default as AnalysisView } from "./AnalysisView";
