+++ apps/extension/src/features/auth/components/index.ts
@@ -1,0 +1,2 @@
+export { default as AuthGuard } from "./AuthGuard";
+export { default as GoogleLoginButton } from "./GoogleLoginButton";
