+++ apps/extension/src/features/layout/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Footer } from "./Footer";
+export { default as NavigationTabs } from "./NavigationTabs";
+export { default as PageContent } from "./PageContent";
+export { default as PageHeader } from "./PageHeader";
