+++ apps/extension/src/features/setting/components/index.ts
@@ -1,0 +1,1 @@
+export { default as SettingView } from "./SettingView";
